DjangoAPI/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
Django==3.2.6
django-cors-headers==3.7.0
djangorestframework==3.12.4
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
wincertstore==0.2
